% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bri_plot.R
\name{bri_plot}
\alias{bri_plot}
\title{Plotar mapa de desigualdade racial}
\usage{
bri_plot(
  estado,
  etnia = c("PRETO", "BRANCO", "PARDO", "INDIGENA", "AMARELO"),
  tipo = c("renda", "pobreza", "pobrezax"),
  fonte = c("censo", "rais"),
  ref = c("2010", "2017", "2000", "1991"),
  n_nomes = 2,
  p_nomes = 1000,
  from = 0,
  to = 3000,
  by = 500,
  bar = TRUE,
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{estado}{String. O estado a ser plotado.}

\item{etnia}{String. As etnias a serem plotadas. Cada etnia gera um
objeto ggplot2 que e composto usando o patchwork.}

\item{tipo}{String. Variavel socioeconomia a ser utilizada para a plotagem.}

\item{fonte}{String. Fonte de dados utilizada. Atualmente disponivel o
Censo do IBGE e a Rais do MTE.}

\item{ref}{String. Ano de referencia da coleta de dados. Atualmente
disponiveis os Censos de 1991, 2000 e 2010.}

\item{n_nomes}{Integer. A quantidade de municipios a serem passados ao
geom_label(). Note que a quantidade refere-se tanto aos
municipios de menor e maior renda, de forma que, no total, serao exibidos
o dobro de nomes informados.}

\item{p_nomes}{Integer. A populaçao minima no municipio de pessoas de cada
etnia selecionada para que o nome do municipio seja exibido
pelo geom_label().}

\item{from, to, by}{Integer. Valores a serem passados a seq() para criar a
colorbar. Recomenda-se utilizar valores que representem os minimos e
maximos exibidos.}

\item{bar}{Logical. Se TRUE, a colorbar da renda e plotada.}

\item{title, subtitle, caption}{String. Controlam o titulo, sibtitulo e
legenda do maior nivel.}
}
\description{
Plota o mapa da desigualdade racial nos municipios brasileiros.
Cada mapa e um objeto ggplot2 combinado pelo \code{patchwork}.
}
\details{
A definicao de pobreza e pobreza extrema para o IBGE sao,
respectivamente, pessoas de renda inferior a 1/2 salario minimo e pessoas
de renda inferior a 1/4 salario minimo.
}
