% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_2010.R
\docType{data}
\name{data_2010}
\alias{data_2010}
\title{Dados de Censo do IBGE}
\format{
A tibble with 5565 rows and 34 variables:
\describe{
\item{abbrev_state}{siglas dos estados}
\item{name_muni}{nome do município}
\item{geom}{limites geográficos dos municípios}
}
}
\source{
\url{http://tabnet.datasus.gov.br/cgi/ibge/censo/rendadescr.htm}
}
\usage{
data_2010
}
\description{
Um dataset contendo dados do último censo do IBGE com informações de raça.
Adicionalmente, foi calculada a renda média e desemprego, e incluída as informações
geográficas dos municípios a partir do pacote geobr.
}
\keyword{datasets}
