% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_m.R
\name{r_m}
\alias{r_m}
\title{Plotar grafico de renda media branca x preta}
\usage{
r_m(
  estado,
  from = 500,
  to = 2500,
  by = 500,
  n_nomes = 5,
  p_nomes = 0,
  etnia = c("PRETO", "BRANCO", "PARDO", "INDIGENA", "AMARELO"),
  autor = NULL,
  titulo = NULL,
  subtitulo = NULL
)
}
\arguments{
\item{estado}{String. O estado a ser plotado.}

\item{from, to, by}{Integer. Valores a serem passados à seq() para criar a
colorbar. Recomenda-se utilizar valores que representem os mínimos e
máximos exibidos.}

\item{n_nomes}{Integer. A quantidade de municípios a serem passados ao
geom_label(). Note que a quantidade refere-se tanto aos
municípios de menor e maior renda, de forma que, no total, serão exibidos
o dobro de nomes informados.}

\item{p_nomes}{Integer. A população mínima no município de pessoas de cada
etnia selecionada para que o nome do município seja exibido
pelo geom_label().}

\item{etnia}{String. As etnias a serem plotadas. Cada etnia gera um
objeto ggplot2 que é composto usando o patchwork.}
}
\description{
Essa função plota gráficos de renda média per capita por etnia.
}
