% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rais_2017.R
\docType{data}
\name{rais_2017}
\alias{rais_2017}
\title{Annual Social Information Report (Rais) 2017}
\format{
A list with 27 elements (states), each with 22 rows and 43 variables:
\describe{
\item{abbrev_state}{siglas dos estados}
\item{name_muni}{nome do municipio}
\item{geom}{limites geograficos dos municípios}
}
}
\source{
\url{ftp://ftp.mtps.gov.br/pdet/microdados/}
}
\usage{
rais_2017
}
\description{
A list of dataframes containing the Rais for the year 2017 of each Brazilian
state. The dataframes were summarized so they are in the same structure as
of Census. Note that Rais's race information is filled in by the employer,
often by outsourced third party, and not by the worker itself. Thus, there is
a substancial amount of undeclared racial status.
}
\keyword{datasets}
