% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RM.R
\name{RM}
\alias{RM}
\title{Plotar gráfico de renda média branca x preta}
\usage{
RM(estado, from = 500, to = 2500, by = 500, n.nomes = 5)
}
\description{
Essa função plota os gráficos de renda média per capita branca e preta.
}
